Aggiunto foreign key allo schema delle timbrature e consentito accesso alle risorse nelle rotte delle timbrature solo se in possesso del token

+++ routes/timbratura.js
@@ -7,7 +7,7 @@ router.get("/", verifyToken, async (req, res) => {
   try {
     const limit = parseInt(req.query.limit || "");
     const sort = req.query.order || "";
-    const elencoTimbrature = await Timbratura.find()
+    const elencoTimbrature = await Timbratura.find({ userId: req.user.id })
       .sort({ ingresso: sort })
       .limit(limit);
     res.json(elencoTimbrature);
@@ -21,19 +21,24 @@ router.post("/", verifyToken, async (req, res) => {
     ingresso: req.body.ingresso,
     uscita: req.body.uscita,
     differenza: req.body.differenza,
+    userId: req.user.id,
   });
+  console.log(timbratura);
   try {
     const result = await timbratura.save();
-    console.log(result);
     res.json(result);
   } catch (err) {
-    res.json({ message: err });
+    res.json({ error: err });
   }
 });
 
 router.get("/:id", verifyToken, async (req, res) => {
   try {
     const timbratura = await Timbratura.findById(req.params.id);
+    if (timbratura.userId != req.user.id)
+      return res
+        .status(401)
+        .json({ error: "Non sei autorizzato a visualizzare questa risorsa" });
     res.json(timbratura);
   } catch (err) {
     res.json({ message: err });
@@ -42,6 +47,12 @@ router.get("/:id", verifyToken, async (req, res) => {
 
 router.put("/:id", verifyToken, async (req, res) => {
   try {
+    const timbratura = await Timbratura.findById(req.params.id);
+    if (timbratura.userId != req.user.id)
+      return res
+        .status(401)
+        .json({ error: "Non sei autorizzato a modificare questa risorsa" });
+
     const timbraturaAggiornata = await Timbratura.updateOne(
       {
         _id: req.params.id,
@@ -59,6 +70,11 @@ router.put("/:id", verifyToken, async (req, res) => {
 });
 
 router.delete("/:id", verifyToken, async (req, res) => {
+  const timbratura = await Timbratura.findById(req.params.id);
+  if (timbratura.userId != req.user.id)
+    return res
+      .status(401)
+      .json({ error: "Non sei autorizzato a visualizzare questa risorsa" });
   try {
     const timbraturaCancellata = await Timbratura.deleteOne({
       _id: req.params.id,
